requirements.txt
swampdragon
django-model-utils
tox
pytest-django==2.8.0
pytest==2.6.4
pytest-cov==1.6
flake8==2.2.2